refactor(act,act-http): move autoclose config to own file + zod-validate openapi
Two cleanups landing together:

## 1. autoclose-config split out of act.ts

Moved the autoclose config plumbing — schema, defaults, type,
resolver — out of `libs/act/src/act.ts` into a new
`libs/act/src/internal/autoclose-config.ts`. The orchestrator file
goes back to being focused on the `Act` class; the autoclose
surface lives next to the cycle and controller that consume it.

- `DEFAULT_AUTOCLOSE_CYCLE_MS` / `DEFAULT_CLOSE_BATCH_SIZE` /
  `DEFAULT_CLOSE_YIELD_MS` constants — moved.
- `AutoCloseSchema` Zod schema — moved.
- `AutoCloseConfig` type — moved.
- `resolve_autoclose_config` function — moved.

`act.ts` keeps a single pass-through re-export block so operators
still import `{ DEFAULT_AUTOCLOSE_CYCLE_MS, resolve_autoclose_config
}` from `@rotorsoft/act` — no consumer breakage. The `ActOptions`
type fields stay in `act.ts` since they're part of the public
`ActOptions` interface.

`internal/index.ts` re-exports the new module so other internal
files (the cycle, the controller) consume it through the barrel
and not via direct file paths.

## 2. openapi `validate_options` → Zod schema

Same pattern as the autoclose / sse refactors in the previous
commits. The OpenAPI emitter's `validate_options` had two
hand-rolled string-non-empty checks (`info.title` / `info.version`)
and a for-loop URL parse over each server entry. Replaced with
`OpenAPIValidationSchema` — a `z.object` with `z.string().min(1)`
plus a `z.array(...)` whose items refine via the existing
`is_valid_server_url` helper (the ReDoS-fix regex stays
unchanged).

Custom error messages preserved so the existing test assertions
(`toThrow(/title is required/)`, etc.) still match without the
tests needing to learn ZodError shapes.

## Coverage

Workspace 100% statements / 100% branches / 100% functions /
100% lines (2313 tests passing). Stability snapshot regenerated
to reflect the moved exports.
